+++ cli/types/tests/plugins-test.ts
@@ -48,7 +48,7 @@ cy.login(Cypress.env('username'), Cypress.env('password'))
 
 // "myLabel" property has been added to "interface cy"
 // thus it is NOT of the command chain.
-// $ExpectError
+// @ts-expect-error
 cy.get('.element').myLabel
 
 // $ExpectType number